Cache correctly scoped dependencies' tgz files
The logic for caching .tgz files does not work for scoped dependencies.
It works by getting name and version from the url. For normal packages urls are:
 - https://registry.npmjs.org/double-ended-queue/-/double-ended-queue-2.1.0-0.tgz
But for scoped dependencies they are:
 - https://registry.npmjs.org/@angular/core/-/core-2.0.0.tgz

Use url.parse to correctly identify the name and version of the package and be able to cache the .tgz forever.