Use control file to populate options object
The module parses command line options in order to modify it's behavior, for example where should be the log file, should the logging be verbose, on which port to start, what's the host, etc.
However parsing them from the command line is difficult to use the same package on different machines with different configurations. So add -c option that will be used to pass a file with configuration.
The file should be just json, containing the supported options. Any option from file will overwrite the option passed on the command line. Example content is:
```JSON
{
	"logPath": "./test1",
	"verbose": true
}

```
This way the logs will go to ./test1 file and the mode will be verbose.

Remove the shell script from .bin directory - we do not need it anymore.
Use lodash as... it's lodash, just use it.