Do not break process when error is raised in cache.meta
In some cases we call the callback passed to cache.meta with error object.
This callback is inside proxy.js's http handler and it automatically raises the error.
This error is unhandled, so the process breaks.

Instead of doing this, just return the error. This way the process should remain alive and the npm client should retry the operation.